get user by query
