+++ apiServer.js
@@ -110,220 +110,20 @@ app.post('/user/changePasswordForgot', userApis.POST_changePasswordForgot);
 app.post('/candidate/sendVerificationEmail', candidateApis.POST_sendVerificationEmail);
 app.post('/candidate/completePathway', candidateApis.POST_completePathway);
 app.post("/candidate/addPathway", candidateApis.POST_addPathway);
+app.post('/candidate/comingSoonEmail', candidateApis.POST_comingSoonEmail);
 
 app.post('/business/forBusinessEmail', businessApis.POST_forBusinessEmail);
+app.post('/business/contactUsEmail', businessApis.POST_contactUsEmail);
 
 app.post("/admin/alertLinkClicked", adminApis.POST_alertLinkClicked);
 
+app.get('/user/userByProfileUrl', userApis.GET_userByProfileUrl);
 app.post('/user/changePassword', userApis.POST_changePassword);
+app.post('/user/forgotPassword', userApis.POST_forgotPassword);
 
 app.post('/misc/createReferralCode', miscApis.POST_createReferralCode);
+app.post('/user/unsubscribeEmail', miscApis.POST_unsubscribeEmail);
 
-
-app.post('/user/unsubscribeEmail', function (req, res) {
-
-    let recipient = ["kyle@moonshotlearning.org"];
-    let subject = 'URGENT ACTION - User Unsubscribe from Moonshot';
-    let content = "<div>"
-        + "<h3>This email is Unsubscribing from Moonshot Emails:</h3>"
-        + "<p>Email: "
-        + sanitize(req.body.email)
-        + "</p>"
-        + "</div>";
-
-    const sendFrom = "Moonshot";
-    sendEmail(recipient, subject, content, sendFrom, undefined, function (success, msg) {
-        if (success) {
-            res.json("You have successfully unsubscribed.");
-        } else {
-            res.status(500).send(msg);
-        }
-    });
-
-    const optOutError = function(error) {
-        console.log("ERROR ADDING EMAIL TO OPT OUT LIST: " + req.body.email);
-        console.log("The error was: ", error);
-        let recipient = ["ameyer24@wisc.edu"];
-        let subject = "MOONSHOT - URGENT ACTION - User was not unsubscribed"
-        let content = "<div>"
-            + "<h3>This email could not be added to the optOut list:</h3>"
-            + "<p>Email: "
-            + sanitize(req.body.email)
-            + "</p>"
-            + "</div>";
-        sendEmail(recipient, subject, content, sendFrom, undefined, function(){});
-    }
-
-    // add email to list of unsubscribed emails
-    Emailaddresses.findOne({name: "optedOut"}, function(err, optedOut) {
-        if (err) {
-            optOutError(err);
-        }
-        else {
-            console.log("adding to opted-out list: ", req.body.email)
-            optedOut.emails.push(req.body.email);
-            optedOut.save(function(err2, newOptedOut) {
-                if (err2) {
-                    optOutError(err2);
-                }
-            });
-        }
-    });
-});
-
-
-// SEND COMING SOON EMAIL
-app.post('/user/comingSoonEmail', function (req, res) {
-
-    let recipient = ["kyle@moonshotlearning.org", "justin@moonshotlearning.org", "ameyer24@wisc.edu"];
-    let subject = 'Moonshot Coming Soon Pathway';
-    let content = "<div>"
-        + "<h3>Pathway:</h3>"
-        + "<p>Name: "
-        + sanitize(req.body.name)
-        + "<p>Email: "
-        + sanitize(req.body.email)
-        + "</p>"
-        + "<p>Pathway: "
-        + sanitize(req.body.pathway)
-        + "</p>"
-        + "</div>";
-
-    const sendFrom = "Moonshot";
-    sendEmail(recipient, subject, content, sendFrom, undefined, function (success, msg) {
-        if (success) {
-            res.json("Email sent successfully, our team will be in contact with you shortly!");
-        } else {
-            res.status(500).send(msg);
-        }
-    })
-});
-
-// SEND EMAIL FOR CONTACT US
-app.post('/user/contactUsEmail', function (req, res) {
-
-    let message = "None";
-    if (req.body.message) {
-        message = sanitize(req.body.message);
-    }
-    let recipients = ["kyle@moonshotlearning.org", "justin@moonshotlearning.org"];
-    let subject = 'Moonshot Pathway Question -- Contact Us Form';
-    let content = "<div>"
-        + "<h3>Questions from pathway:</h3>"
-        + "<p>Name: "
-        + sanitize(req.body.name)
-        + "</p>"
-        + "<p>Email: "
-        + sanitize(req.body.email)
-        + "</p>"
-        + "<p>Message: "
-        + message
-        + "</p>"
-        + "</div>";
-
-    const sendFrom = "Moonshot";
-    sendEmail(recipients, subject, content, sendFrom, undefined, function (success, msg) {
-        if (success) {
-            res.json("Email sent successfully, our team will be in contact with you shortly!");
-        } else {
-            res.status(500).send(msg);
-        }
-    })
-});
-
-// SEND EMAIL FOR PASSWORD RESET
-app.post('/forgotPassword', function (req, res) {
-
-    let email = sanitize(req.body.email);
-    let query = {email: email};
-
-    const user = getUserByQuery(query, function (err, user) {
-        if (!user) {
-            console.log("Couldn't find user to set their password change token.");
-            res.status(401).send("Cannot find user");
-            return;
-        } else {
-            // token that will go in the url
-            const newPasswordToken = crypto.randomBytes(64).toString('hex');
-            // password token expires in one hour (minutes * seconds * milliseconds)
-            const newTime = Date.now() + (60 * 60 * 1000);
-
-            const query2 = {_id: user._id};
-            const update = {
-                '$set': {
-                    passwordToken: newPasswordToken,
-                    passwordTokenExpirationTime: newTime,
-                }
-            };
-
-            const options = {new: true};
-
-            Users.findOneAndUpdate(query2, update, options, function (err, foundUser) {
-                if (err) {
-                    console.log("Error giving user reset-password token: ", err);
-                    res.status(500).send("Server error, try again later.");
-                    return;
-                }
-
-                // if we're in development (on localhost) navigate to localhost
-                let moonshotUrl = "https://www.moonshotlearning.org/";
-                if (!process.env.NODE_ENV) {
-                    moonshotUrl = "http://localhost:8081/";
-                }
-                const recipient = [user.email];
-                const subject = 'Change Password';
-
-                const content =
-                    '<div style="font-size:15px;text-align:center;font-family: Arial, sans-serif;color:#686868">'
-                        + '<a href="' + moonshotUrl + '" style="color:#00c3ff"><img alt="Moonshot Logo" style="height:100px;margin-bottom:20px"src="https://image.ibb.co/iAchLn/Official_Logo_Blue.png"/></a><br/>'
-                            + '<div style="text-align:justify;width:80%;margin-left:10%;">'
-                            + "<span style='margin-bottom:20px;display:inline-block;'>Hello! We got a request to change your password. If that wasn't from you, you can ignore this email and your password will stay the same. Otherwise click here:</span><br/>"
-                            + '</div>'
-                        + '<a style="display:inline-block;height:28px;width:170px;font-size:18px;border:2px solid #00d2ff;color:#00d2ff;padding:10px 5px 0px;text-decoration:none;margin:5px 20px 20px;" href="' + moonshotUrl + 'changePassword?token='
-                        + newPasswordToken
-                        + '">Change Password</a>'
-                        + '<div style="text-align:left;width:80%;margin-left:10%;">'
-                            + '<div style="font-size:10px; text-align:center; color:#C8C8C8; margin-bottom:30px;">'
-                                + '<i>Moonshot Learning, Inc.<br/><a href="" style="text-decoration:none;color:#D8D8D8;">1261 Meadow Sweet Dr<br/>Madison, WI 53719</a>.<br/>'
-                                + '<a style="color:#C8C8C8; margin-top:20px;" href="' + moonshotUrl + 'unsubscribe?email=' + user.email + '">Opt-out of future messages.</a></i>'
-                            + '</div>'
-                        + '</div>'
-                    + '</div>';
-
-                const sendFrom = "Moonshot";
-                sendEmail(recipient, subject, content, sendFrom, undefined, function (success, msg) {
-                    if (success) {
-                        res.json(msg);
-                    } else {
-                        res.status(500).send(msg);
-                    }
-                })
-            });
-        }
-    })
-});
-
-
-
-app.post('/getUserById', function (req, res) {
-    const _id = sanitize(req.body._id);
-    const query = {_id};
-    getUserByQuery(query, function (err, user) {
-        if (err) {
-            res.status(500).send("User not found.");
-        } else {
-            res.json(safeUser(user));
-        }
-    })
-});
-
-app.post('/getUserByProfileUrl', function (req, res) {
-    const profileUrl = sanitize(req.body.profileUrl);
-    const query = { profileUrl };
-    getUserByQuery(query, function (err, user) {
-        res.json(safeUser(user));
-    })
-});
 
 
 // LOGIN USER
