+++ apis/businessApis.js
@@ -16,7 +16,8 @@ const { sanitize,
 
 
 const businessApis = {
-    POST_forBusinessEmail
+    POST_forBusinessEmail,
+    POST_contactUsEmail
 }
 
 
@@ -56,6 +57,37 @@ function POST_forBusinessEmail(req, res) {
     sendEmail(recipients, subject, content, sendFrom, undefined, function (success, msg) {
         if (success) {
             res.json("Email sent successfully, our team will notify you of your results shortly.");
+        } else {
+            res.status(500).send(msg);
+        }
+    })
+}
+
+
+function POST_contactUsEmail(req, res) {
+    let message = "None";
+    if (req.body.message) {
+        message = sanitize(req.body.message);
+    }
+    let recipients = ["kyle@moonshotlearning.org", "justin@moonshotlearning.org"];
+    let subject = 'Moonshot Pathway Question -- Contact Us Form';
+    let content = "<div>"
+        + "<h3>Questions from pathway:</h3>"
+        + "<p>Name: "
+        + sanitize(req.body.name)
+        + "</p>"
+        + "<p>Email: "
+        + sanitize(req.body.email)
+        + "</p>"
+        + "<p>Message: "
+        + message
+        + "</p>"
+        + "</div>";
+
+    const sendFrom = "Moonshot";
+    sendEmail(recipients, subject, content, sendFrom, undefined, function (success, msg) {
+        if (success) {
+            res.json("Email sent successfully, our team will be in contact with you shortly!");
         } else {
             res.status(500).send(msg);
         }
