+++ src/components/pages/businessPages/businessProfile.js
@@ -45,8 +45,8 @@ class BusinessProfile extends Component {
 
         // looking at someone else's profile
         else {
-            axios.post("/api/getUserByProfileUrl", {profileUrl}
-            ).then(res => {
+            axios.get("/api/user/userByProfileUrl", {params: {profileUrl}})
+            .then(res => {
                 const user = res.data;
 
                 this.setState({
